// src/app/shared/enums/messages.enum.ts
export enum Messages {
    // Generic Messages
    ERROR = 'Ocorreu um erro.',

    // Titles
    //
    // Device
    DEVICE_TITLE_NEW = 'Novo Device.',
    DEVICE_TITLE_UPDATE = 'Atualizar Device.',
    DEVICE_TITLE_DELETE = 'Excluir Device.',

    // User
    USER_TITLE_NEW = 'Novo Usuário.',
    USER_TITLE_UPDATE = 'Atualizar Usuário.',
    USER_TITLE_DELETE = 'Excluir Usuário.',

    // Messages
    //
    // Device
    DEVICE_INSERT = 'Device salvo com sucesso.',
    DEVICE_INSERT_ERROR = 'Erro ao salvar device.',
    DEVICE_UPDATE = 'Device atualizado com sucesso.',
    DEVICE_UPDATE_ERROR = 'Erro ao atualizar device',
    DEVICE_DELETE = 'Device excluído com sucesso.',
    DEVICE_DELETE_ERROR = 'Erro ao excluir device.',

    // User
    USER_INSERT = 'Usuário salvo com sucesso.',
    USER_INSERT_ERROR = 'Erro ao salvar usuário.',
    USER_UPDATE = 'Usuário atualizado com sucesso.',
    USER_UPDATE_ERROR = 'Erro ao atualizar usuário',
    USER_DELETE = 'Usuário excluído com sucesso.',
    USER_DELETE_ERROR = 'Erro ao excluir usuário.',

    // Tracking
    TRACKING_GET_ERROR = 'Erro ao buscar rastreamento.'
}